pyproject.toml
[tool.poetry]
name = "nonebot-plugin-qqmusic-reco"
version = "0.1.9"
description = "基于QQ音乐歌单，支持多群配置、持久化管理及定时自定义话术的音乐推荐插件"
authors = ["ChlorophyTeio <chlorophyimo@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/ChlorophyTeio/nonebot-plugin-qqmusic-reco"
repository = "https://github.com/ChlorophyTeio/nonebot-plugin-qqmusic-reco"
keywords = ["nonebot", "nonebot2", "qqmusic", "music", "recommendation"]
packages = [
    { include = "nonebot_plugin_qqmusic_reco" }
]

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.3.0"
nonebot-adapter-onebot = "^2.2.0"
nonebot-plugin-apscheduler = "^0.4.0"
nonebot-plugin-localstore = "^0.6.0"
httpx = ">=0.23.0"
pydantic = ">=1.10.0,<3.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.1.0"
isort = "^5.13.0"
nonebug = "^0.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"